修正不上传头像错误问题

+++ server/controllers/signUp.js
@@ -5,7 +5,7 @@ const formidable = require("formidable");
 const uuidV4 = require('uuid/v4');
 const path = require('path');
 let UserModel = require('../models/users');
-
+const DEFAUL_AVATAR =  "/avater/spinner.gif"
 function formidablePromise(req, opts) {
 	return new Promise(function(resolve, reject) {
 		var form = new formidable.IncomingForm(opts)
@@ -30,8 +30,24 @@ module.exports = {
 	},
 	//注册api
 	'POST /api/signUp': async(ctx, next) => {
+		//初始化变量
+		let formidableResult,
+		avater=DEFAUL_AVATAR,
+		currentUser,
+		code = 1,
+		message = "注册成功",
+		fileName,
+		dataBuffer,
+		base64Data;
 		try {
-			var formidableResult = await formidablePromise(ctx.req);
+			formidableResult = await formidablePromise(ctx.req),
+			avater=DEFAUL_AVATAR,
+			currentUser,
+			code = 1,
+			message = "注册成功",
+			fileName,
+			dataBuffer,
+			base64Data;
 			//从fromdata中取出数据
 			let {
 				account,
@@ -41,18 +57,20 @@ module.exports = {
 				password,
 				pic
 			} = formidableResult.fields;
-			//正则替换base64
-			let base64Data = pic.replace(/^data:image\/\w+;base64,/, "");
-			//转换为数据流
-			let dataBuffer = new Buffer(base64Data, 'base64');
-			//生成文件名
-			var fileName = uuidV4().replace(/-/g, ""),
-				code = "1",
-				message = "注册成功";
-			//生成绝对路径
-			var avatarPath = path.resolve(__dirname, "../"),
-				//组合出本地存储图片的路径
-				avater = path.join(avatarPath, `upload/avater/${fileName}.png`);
+			if(pic){
+				//有上传头像...
+       //正则替换base64
+				 base64Data = pic.replace(/^data:image\/\w+;base64,/, "");
+				//转换为数据流
+				 dataBuffer = new Buffer(base64Data, 'base64');
+				//生成文件名
+				  fileName = uuidV4().replace(/-/g, "");
+				  //生成绝对路径
+				   avatarPath = path.resolve(__dirname, "../"),
+					//组合出本地存储图片的路径
+					avater = path.join(avatarPath, `upload/avater/${fileName}.png`);
+			}
+
 			password = crypto.createHash('md5').update(password).digest('hex');
 			let user = {
 				account: account,
@@ -60,14 +78,22 @@ module.exports = {
 				password: password,
 				gender: gender,
 				bio: bio,
-				avatar: `/avater/${fileName}.png`
+				avatar: avater
 			};
+
+			//默认头像用户不需要生成头像
+			if(avater == DEFAUL_AVATAR){
+				 result = await UserModel.create(user)
+				 currentUser = result.ops[0];
+			}else{
+				result = await Promise.all([fs.writeFile(avater, dataBuffer), UserModel.create(user)])
+				currentUser = result[1].ops[0];
+			}
 			//保存图片和生成用户
-			result = await Promise.all([fs.writeFile(avater, dataBuffer), UserModel.create(user)])
-			var currentUser = result[1].ops[0];
+
 			delete user.password;
-			ctx.session.user = user;
 		} catch (e) {
+			console.log(e)
 			if (e.message.match('E11000 duplicate key')) {
 				code = -1;
 				message = "用户名已被占用"
@@ -75,10 +101,11 @@ module.exports = {
 				code = -2;
 				message = e.message
 			}
-
-			let deletePath = path.resolve(__dirname, "../", avater);
-			//注册失败删除头像
-			await fs.unlink(deletePath);
+     if(avater != DEFAUL_AVATAR){
+			 let deletePath = path.resolve(__dirname, "../", avater);
+ 			//注册失败删除头像
+ 			await fs.unlink(deletePath);
+		 }
 		}
 		ctx.response.body = {
 			"code": code,
